New func

+++ ToDoTaskManager.Application/ToDoService.cs
@@ -3,16 +3,19 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using ToDoTaskManager.Domain.Core;
 using ToDoTaskManager.Domain.ToDos;
 namespace ToDoTaskManager.Application;
 
 public class ToDoService
 {
     private readonly IToDoRepository _toDoRepository;
+    private readonly IUnitOfWork _unitOfWork;
 
-    public ToDoService(IToDoRepository toDoRepository)
+    public ToDoService(IToDoRepository toDoRepository, IUnitOfWork unitOfWork)
     {
         _toDoRepository = toDoRepository;
+        _unitOfWork = unitOfWork;
     }
 
     public async Task<Guid> CreateNewToDoAsync(string name, CancellationToken cancellationToken = default)
@@ -26,7 +29,7 @@ public class ToDoService
 
         await _toDoRepository.AddAsync(toDo, cancellationToken);
 
-        await _toDoRepository.SaveChangesAsync(cancellationToken);
+        await _unitOfWork.SaveChangesAsync(cancellationToken);
 
         return toDo.Id;
     }
@@ -35,7 +38,12 @@ public class ToDoService
     {
         var toDo = await _toDoRepository.GetByIdAsync(id, cancellationToken);
 
-        await _toDoRepository.SaveChangesAsync(cancellationToken);
+        return toDo;
+    }
+
+    public async Task<IEnumerable<ToDo>> GetToDos(int count, int page, CancellationToken cancellationToken = default) 
+    {
+        var toDo = await _toDoRepository.GetToDos(count, page, cancellationToken);
 
         return toDo;
     }
@@ -43,9 +51,30 @@ public class ToDoService
     public async Task DeleteById(Guid id, CancellationToken cancellationToken = default) 
     {
         var toDo = await _toDoRepository.GetByIdAsync(id, cancellationToken);
+
+        if (toDo == null) 
+        {
+            throw new ToDoNotFoundExeption(id);
+        }
         
         await _toDoRepository.DeleteAsync(toDo, cancellationToken);
 
-        await _toDoRepository.SaveChangesAsync(cancellationToken);
+        await _unitOfWork.SaveChangesAsync(cancellationToken);
+    }
+
+    public async Task CloseToDoById(Guid id, CancellationToken cancellationToken = default) 
+    {
+        var toDo = await _toDoRepository.GetByIdAsync(id, cancellationToken);
+
+        if (toDo == null) 
+        {
+            throw new ToDoNotFoundExeption(id);
+        }
+
+        toDo.Close();
+
+        await _toDoRepository.UpdateAsync(toDo, cancellationToken);
+
+        await _unitOfWork.SaveChangesAsync(cancellationToken);
     }
 }
